+++ CHANGES.rst
@@ -18,6 +18,8 @@ Changelog
 - @types improvements: add `widget`, support RichText field
   [ebrehault]
 
+- add fieldsets in @types
+  [ebrehault]
 
 1.0a1 (2016-07-14)
 ------------------